Revert "scheduler: task.Data made 64bit to avoid overflow"
This reverts commit fd8938c6346010979992986b2a1b512449e5c866.

+++ src/internal/task/task.go
@@ -13,7 +13,7 @@ type Task struct {
 	Ptr unsafe.Pointer
 
 	// Data is a field which can be used for storing state information.
-	Data uint64
+	Data uint
 
 	// state is the underlying running state of the task.
 	state state
